refactor: 이벤트 리스너 액티비티의 private class로
